Narrow MAX_SPLITS and MAX_MINTS from usize to u8
usize is not IDL-compatible, and these values are small enough
for u8. Adjust call sites to cast where needed.

+++ programs/flex/src/instructions/deposit.rs
@@ -64,7 +64,7 @@ pub fn deposit(ctx: Context<Deposit>, amount: u64) -> Result<()> {
             FlexError::OwnerOnly
         );
         require!(
-            (escrow.mint_count as usize) < MAX_MINTS,
+            escrow.mint_count < MAX_MINTS as u64,
             FlexError::MintLimitReached
         );
